% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions_actions.R
\name{Action}
\alias{Action}
\title{Debugger Action Class}
\description{
Base class for action, which is to be invoked when a rule fires.
             Offers `serialize` function to convert action parameters to a
             string dictionary.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Action$new()}}
\item \href{#method-serialize}{\code{Action$serialize()}}
\item \href{#method-format}{\code{Action$format()}}
\item \href{#method-clone}{\code{Action$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
This class is not meant to be initialized directly.
             Accepts dictionary of action parameters and drops keys whose values are `None`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Action$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{: Dictionary of action parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serialize"></a>}}
\if{latex}{\out{\hypertarget{method-serialize}{}}}
\subsection{Method \code{serialize()}}{
Serialize the action parameters as a string dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Action$serialize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Action parameters serialized as a string dictionary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Action$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Action$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
