% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions_actions.R
\name{ActionList}
\alias{ActionList}
\title{Debugger ActionList Action Class}
\description{
Higher level object to maintain a list of actions to be invoked
             when a rule is fired.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ActionList$new()}}
\item \href{#method-update_training_job_prefix_if_specified}{\code{ActionList$update_training_job_prefix_if_specified()}}
\item \href{#method-serialize}{\code{ActionList$serialize()}}
\item \href{#method-clone}{\code{ActionList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Offers higher level `serialize` function to
             handle serialization of actions as a string list of dictionaries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActionList$new(actions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{actions}}{: List of actions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_training_job_prefix_if_specified"></a>}}
\if{latex}{\out{\hypertarget{method-update_training_job_prefix_if_specified}{}}}
\subsection{Method \code{update_training_job_prefix_if_specified()}}{
For any StopTraining actions in the action list, update the trainingjob
             prefix to be the training job name if the user has not already specified
             a custom training job prefix. This is meant to be called via the sagemaker
             SDK when `estimator.fit` is called by the user. Validation is purposely
             excluded here so that any failures in validation of the training job name
             are intentionally caught in the sagemaker SDK and not here.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActionList$update_training_job_prefix_if_specified(training_job_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{training_job_name}}{: Name of the training job, passed in when `estimator.fit` is called.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serialize"></a>}}
\if{latex}{\out{\hypertarget{method-serialize}{}}}
\subsection{Method \code{serialize()}}{
Serialize the action parameters as a string dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActionList$serialize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Action parameters serialized as a string dictionary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActionList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
