% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler_rules_rules.R
\name{BatchSize}
\alias{BatchSize}
\title{Debugger BatchSize class}
\description{
This rule helps to detect if GPU is underulitized because of the batch size being too small.
             To detect this the rule analyzes the average GPU memory footprint, CPU and GPU utilization.
             If utilization on CPU, GPU and memory footprint is on average low , it may indicate that user
             can either run on a smaller instance type or that batch size could be increased. This analysis does not
             work for frameworks that heavily over-allocate memory. Increasing batch size could potentially lead to
             a processing/dataloading bottleneck, because more data needs to be pre-processed in each iteration.
}
\section{Super class}{
\code{\link[R6sagemaker.debugger:ProfilerRuleBase]{R6sagemaker.debugger::ProfilerRuleBase}} -> \code{BatchSize}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BatchSize$new()}}
\item \href{#method-clone}{\code{BatchSize$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize BatchSize class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchSize$new(
  cpu_threshold_p95 = 70,
  gpu_threshold_p95 = 70,
  gpu_memory_threshold_p95 = 70,
  patience = 1000,
  window = 500,
  scan_interval_us = 60 * 1000 * 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cpu_threshold_p95}}{(numeric): defines the threshold for 95th quantile of CPU utilization.Default is 70\%.}

\item{\code{gpu_threshold_p95}}{(numeric): defines the threshold for 95th quantile of GPU utilization.Default is 70\%.}

\item{\code{gpu_memory_threshold_p95}}{(numeric): defines the threshold for 95th quantile of GPU memory utilization.Default is 70\%.}

\item{\code{patience}}{(numeric): defines how many data points to capture before Rule runs the first evluation. Default 100}

\item{\code{window}}{(numeric): window size for computing quantiles.}

\item{\code{scan_interval_us}}{(numeric): interval with which timeline files are scanned. Default is 60000000 us.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchSize$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
