% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler_rules_rules.R
\name{CPUBottleneck}
\alias{CPUBottleneck}
\title{Debugger CPUBottleneck class}
\description{
This rule helps to detect if GPU is underutilized due to CPU bottlenecks.
             Rule returns True if number of CPU bottlenecks exceeds a predefined threshold.
}
\section{Super class}{
\code{\link[R6sagemaker.debugger:ProfilerRuleBase]{R6sagemaker.debugger::ProfilerRuleBase}} -> \code{CPUBottleneck}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CPUBottleneck$new()}}
\item \href{#method-clone}{\code{CPUBottleneck$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize CPUBottleneck class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPUBottleneck$new(
  threshold = 50,
  gpu_threshold = 10,
  cpu_threshold = 90,
  patience = 1000,
  scan_interval_us = 60 * 1000 * 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{threshold}}{: defines the threshold beyond which Rule should return True. Default is 50 percent.
So if there is a bottleneck more than ``50%`` of the time during the training Rule will return True.}

\item{\code{gpu_threshold}}{: threshold that defines when GPU is considered being under-utilized. Default is ``10%``}

\item{\code{cpu_threshold}}{: threshold that defines high CPU utilization. Default is above ``90%``}

\item{\code{patience}}{: How many values to record before checking for CPU bottlenecks. During training
initialization, GPU is likely at 0 percent, so Rule should not check for
under utilization immediately. Default 1000.}

\item{\code{scan_interval_us}}{: interval with which timeline files are scanned. Default is 60000000 us.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPUBottleneck$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
