% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions_actions.R
\name{Email}
\alias{Email}
\title{Debugger Email Action class}
\description{
Action for sending an email to the provided email address when
             the rule is fired. Note that a policy must be created in the AWS
             account to allow the sagemaker role to send an email to the user.
}
\section{Super class}{
\code{\link[sagemaker.debugger:Action]{sagemaker.debugger::Action}} -> \code{Email}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Email$new()}}
\item \href{#method-clone}{\code{Email$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.debugger" data-topic="Action" data-id="format">}\href{../../sagemaker.debugger/html/Action.html#method-format}{\code{sagemaker.debugger::Action$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.debugger" data-topic="Action" data-id="serialize">}\href{../../sagemaker.debugger/html/Action.html#method-serialize}{\code{sagemaker.debugger::Action$serialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize Email action class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Email$new(email_address)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email_address}}{: Email address to send the email notification to.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Email$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
