% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler_rules_rules.R
\name{GPUMemoryIncrease}
\alias{GPUMemoryIncrease}
\title{Debugger GPUMemoryIncrease class}
\description{
This rule helps to detect large increase in memory usage on GPUs.
             The rule computes the moving average of continous datapoints and compares
             it against the moving average of previous iteration.
}
\section{Super class}{
\code{\link[R6sagemaker.debugger:ProfilerRuleBase]{R6sagemaker.debugger::ProfilerRuleBase}} -> \code{GPUMemoryIncrease}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GPUMemoryIncrease$new()}}
\item \href{#method-clone}{\code{GPUMemoryIncrease$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.debugger" data-topic="ProfilerRuleBase" data-id="format">}\href{../../R6sagemaker.debugger/html/ProfilerRuleBase.html#method-format}{\code{R6sagemaker.debugger::ProfilerRuleBase$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize GPUMemoryIncrease class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GPUMemoryIncrease$new(
  increase = 5,
  patience = 1000,
  window = 10,
  scan_interval_us = 60 * 1000 * 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{increase}}{: defines the threshold for absolute memory increase.Default is 5\%.
So if moving average increase from 5\% to 6\%, the rule will fire.}

\item{\code{patience}}{: defines how many continous datapoints to capture before Rule runs
the first evluation. Default is 1000}

\item{\code{window}}{: window size for computing moving average of continous datapoints}

\item{\code{scan_interval_us}}{: interval with which timeline files are scanned. Default is 60000000 us.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GPUMemoryIncrease$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
