% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler_rules_rules.R
\name{IOBottleneck}
\alias{IOBottleneck}
\title{Debugger IOBottleneck class}
\description{
This rule helps to detect if GPU is underutilized due to IO bottlenecks.
             Rule returns True if number of IO bottlenecks exceeds a predefined threshold.
}
\section{Super class}{
\code{\link[R6sagemaker.debugger:ProfilerRuleBase]{R6sagemaker.debugger::ProfilerRuleBase}} -> \code{IOBottleneck}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{IOBottleneck$new()}}
\item \href{#method-clone}{\code{IOBottleneck$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize IOBottleneck class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IOBottleneck$new(
  threshold = 50,
  gpu_threshold = 10,
  io_threshold = 50,
  patience = 1000,
  scan_interval_us = 60 * 1000 * 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{threshold}}{: defines the threshold when Rule should return True. Default
is 50 percent. So if there is a bottleneck more than 50\% of the
time during the training Rule will return True.}

\item{\code{gpu_threshold}}{: threshold that defines when GPU is considered being under-utilized.
Default is 70\%}

\item{\code{io_threshold}}{: threshold that defines high IO wait time. Default is above 50\%}

\item{\code{patience}}{: How many values to record before checking for IO bottlenecks. During
training initilization, GPU is likely at 0 percent, so Rule should not check
for underutilization immediatly. Default 1000.}

\item{\code{scan_interval_us}}{: interval with which timeline files are scanned. Default is 60000000 us.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IOBottleneck$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
