% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler_rules_rules.R
\name{LowGPUUtilization}
\alias{LowGPUUtilization}
\title{Debugger LowGPUUtilization class}
\description{
This rule helps to detect if GPU utilization is low or suffers from
             fluctuations. This is checked for each single GPU on each worker node.
             Rule returns True if 95th quantile is below threshold_p95 which indicates
             under-utilization. Rule returns true if 95th quantile is above threshold_p95
             and 5th quantile is below threshold_p5 which indicates fluctuations.
}
\section{Super class}{
\code{\link[R6sagemaker.debugger:ProfilerRuleBase]{R6sagemaker.debugger::ProfilerRuleBase}} -> \code{LowGPUUtilization}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LowGPUUtilization$new()}}
\item \href{#method-clone}{\code{LowGPUUtilization$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.debugger" data-topic="ProfilerRuleBase" data-id="format">}\href{../../R6sagemaker.debugger/html/ProfilerRuleBase.html#method-format}{\code{R6sagemaker.debugger::ProfilerRuleBase$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize LowGPUUtilization class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LowGPUUtilization$new(
  threshold_p95 = 70,
  threshold_p5 = 10,
  window = 500,
  patience = 1000,
  scan_interval_us = 60 * 1000 * 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{threshold_p95}}{: threshold for 95th quantile below which GPU is considered
to be underutilized. Default is 70 percent.}

\item{\code{threshold_p5}}{: threshold for 5th quantile. Default is 10 percent.}

\item{\code{window}}{: number of past datapoints which are used to compute the quantiles.}

\item{\code{patience}}{: How many values to record before checking for underutilization/fluctuations.
During training initilization, GPU is likely at 0 percent, so Rule should not
check for underutilization immediately. Default 1000.}

\item{\code{scan_interval_us}}{: interval with which timeline files are scanned. Default is 60000000 us.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LowGPUUtilization$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
