% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler_rules_rules.R
\name{MaxInitializationTime}
\alias{MaxInitializationTime}
\title{Debugger MaxInitializationTime class}
\description{
This rule helps to detect if the training intialization is taking too
             much time. The rule waits until first step is available.
}
\section{Super class}{
\code{\link[R6sagemaker.debugger:ProfilerRuleBase]{R6sagemaker.debugger::ProfilerRuleBase}} -> \code{MaxInitializationTime}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MaxInitializationTime$new()}}
\item \href{#method-clone}{\code{MaxInitializationTime$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize MaxInitializationTime class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MaxInitializationTime$new(threshold = 20, scan_interval_us = 60 * 1000 * 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{threshold}}{: defines the threshold in minutes to wait for first step to
become available. Default is 20 minutes.}

\item{\code{scan_interval_us}}{: interval with which timeline files are scanned. Default is 60000000 us.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MaxInitializationTime$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
