% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler_rules_rules.R
\name{OverallSystemUsage}
\alias{OverallSystemUsage}
\title{Debugger OverallSystemUsage class}
\description{
This rule measures overall system usage per worker node. The rule
             currently only aggregates values per node and computes their percentiles. The
             rule does currently not take any threshold parameters into account nor can it trigger.
             The reason behind that is that other rules already cover cases such as under utilization
             and they do it at a more fine-grained level e.g. per GPU. We may change this in the future.
}
\section{Super class}{
\code{\link[R6sagemaker.debugger:ProfilerRuleBase]{R6sagemaker.debugger::ProfilerRuleBase}} -> \code{OverallSystemUsage}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OverallSystemUsage$new()}}
\item \href{#method-clone}{\code{OverallSystemUsage$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.debugger" data-topic="ProfilerRuleBase" data-id="format">}\href{../../R6sagemaker.debugger/html/ProfilerRuleBase.html#method-format}{\code{R6sagemaker.debugger::ProfilerRuleBase$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize OverallSystemUsage class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OverallSystemUsage$new(scan_interval_us = 60 * 1000 * 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scan_interval_us}}{: interval with which timeline files are scanned. Default is 60000000 us.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OverallSystemUsage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
