% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler_rules_rules.R
\name{ProfilerReport}
\alias{ProfilerReport}
\title{Debugger ProfilerReport class}
\description{
This rule will create a profiler report after invoking all of the rules. The parameters
             used in any of these rules can be customized by following this naming scheme:
             <rule_name>_<parameter_name> : value
             Validation is also done here to ensure that:
             \itemize{
               \item{The key names follow the above format}
               \item{rule_name corresponds to a valid rule name.}
               \item{parameter_name corresponds to a valid parameter of this rule.}
               \item{The parameter for this rule's parameter is valid.}
             }
}
\section{Super class}{
\code{\link[sagemaker.debugger:ProfilerRuleBase]{sagemaker.debugger::ProfilerRuleBase}} -> \code{ProfilerReport}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProfilerReport$new()}}
\item \href{#method-clone}{\code{ProfilerReport$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.debugger" data-topic="ProfilerRuleBase" data-id="format">}\href{../../sagemaker.debugger/html/ProfilerRuleBase.html#method-format}{\code{sagemaker.debugger::ProfilerRuleBase$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize ProfilerReport class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfilerReport$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{: Dictionary mapping rule + parameter name to value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProfilerReport$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
