% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{R6sagemaker.debugger-package}
\alias{R6sagemaker.debugger}
\alias{R6sagemaker.debugger-package}
\title{r6 sagemaker: this is just a placeholder}
\description{
`R6sagemaker` debugger for sagemaker operations.
}
\author{
\strong{Maintainer}: Dyfan Jones \email{dyfan.r.jones@gmail.com}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
