% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions_actions.R
\name{SMS}
\alias{SMS}
\title{Debugger SMS Action Class}
\description{
Action for sending an SMS to the provided phone number when the rule
             is fired. Note that a policy must be created in the AWS account to allow
             the sagemaker role to send an SMS to the user.
}
\section{Super class}{
\code{\link[R6sagemaker.debugger:Action]{R6sagemaker.debugger::Action}} -> \code{SMS}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SMS$new()}}
\item \href{#method-clone}{\code{SMS$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.debugger" data-topic="Action" data-id="format">}\href{../../R6sagemaker.debugger/html/Action.html#method-format}{\code{R6sagemaker.debugger::Action$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.debugger" data-topic="Action" data-id="serialize">}\href{../../R6sagemaker.debugger/html/Action.html#method-serialize}{\code{R6sagemaker.debugger::Action$serialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize SMS action class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMS$new(phone_number)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phone_number}}{: Valid phone number that follows the the E.164 format. See
\url{https://docs.aws.amazon.com/sns/latest/dg/sms_publish-to-phone.html}
for more info.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
