% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions_actions.R
\name{StopTraining}
\alias{StopTraining}
\title{Debugger StopTraining Action class}
\description{
Action for stopping the training job when a rule is fired.
}
\section{Super class}{
\code{\link[R6sagemaker.debugger:Action]{R6sagemaker.debugger::Action}} -> \code{StopTraining}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StopTraining$new()}}
\item \href{#method-update_training_job_prefix_if_not_specified}{\code{StopTraining$update_training_job_prefix_if_not_specified()}}
\item \href{#method-clone}{\code{StopTraining$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.debugger" data-topic="Action" data-id="format">}\href{../../R6sagemaker.debugger/html/Action.html#method-format}{\code{R6sagemaker.debugger::Action$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="R6sagemaker.debugger" data-topic="Action" data-id="serialize">}\href{../../R6sagemaker.debugger/html/Action.html#method-serialize}{\code{R6sagemaker.debugger::Action$serialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Note that a policy must be created in the AWS account to allow the sagemaker
             role to stop the training job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StopTraining$new(training_job_prefix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{training_job_prefix}}{: The prefix of the training job to stop if the rule is fired.
This must only refer to one active training job, otherwise no training job will
be stopped.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_training_job_prefix_if_not_specified"></a>}}
\if{latex}{\out{\hypertarget{method-update_training_job_prefix_if_not_specified}{}}}
\subsection{Method \code{update_training_job_prefix_if_not_specified()}}{
Update the training job prefix to be the training job name if the user
             has not already specified a custom training job prefix. This is only meant
             to be called via the sagemaker SDK when `estimator.fit` is called by the
             user. Validation is purposely excluded here so that any failures in validation
             of the training job name are intentionally caught in the sagemaker SDK and not here.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StopTraining$update_training_job_prefix_if_not_specified(training_job_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{training_job_name}}{: Name of the training job, passed in when `estimator.fit` is called.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StopTraining$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
